Fix restoring video playback state on configuration change
Set isPlayerInitialised as true in playVideo():
Without these we would incorrectly initialise the preview instead of the player when
the activity is resumed after onStop.

Get bundle in onPause/onStop rather than onDestroy:
onSaveInstanceState() is called around the time onStop is called, so getting the bundle
in onDestroy would be too late.